Replace lsblk fs detection with blkid
